requirements.txt
openai
sentence-transformers
faiss-cpu
python-dotenv
redis
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.7
requests==2.31.0
streamlit==1.25.0
aiofiles==23.1.0
numpy
scipy
nltk
pytest
